+++ ricsdl-package/ricsdl/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2019 AT&T Intellectual Property.
-# Copyright (c) 2018-2019 Nokia.
+# Copyright (c) 2018-2022 Nokia.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -31,7 +31,7 @@ from .exceptions import (
 )
 
 
-__version__ = '3.0.2'
+__version__ = '3.1.0'
 
 
 __all__ = [